# CoinStore benchmark and performance improvements