--- a/chia/clvm/spend_sim.py
+++ b/chia/clvm/spend_sim.py
@@ -129,8 +129,9 @@ class SpendSim:
             uint64(calculate_base_farmer_reward(next_block_height) + fees),
             self.defaults.GENESIS_CHALLENGE,
         )
-        await self.mempool_manager.coin_store._add_coin_record(self.new_coin_record(pool_coin, True))
-        await self.mempool_manager.coin_store._add_coin_record(self.new_coin_record(farmer_coin, True))
+        await self.mempool_manager.coin_store._add_coin_records(
+            [self.new_coin_record(pool_coin, True), self.new_coin_record(farmer_coin, True)]
+        )
 
         # Coin store gets updated
         generator_bundle: Optional[SpendBundle] = None
@@ -147,10 +148,12 @@ class SpendSim:
                     return_additions = additions
                     return_removals = removals
 
-                for addition in additions:
-                    await self.mempool_manager.coin_store._add_coin_record(self.new_coin_record(addition))
-                for removal in removals:
-                    await self.mempool_manager.coin_store._set_spent(removal.name(), uint32(self.block_height + 1))
+                await self.mempool_manager.coin_store._add_coin_records(
+                    [self.new_coin_record(addition) for addition in additions]
+                )
+                await self.mempool_manager.coin_store._set_spent(
+                    [r.name() for r in removals], uint32(self.block_height + 1)
+                )
 
         # SimBlockRecord is created
         generator: Optional[BlockGenerator] = await self.generate_transaction_generator(generator_bundle)