--- a/chia/full_node/coin_store.py
+++ b/chia/full_node/coin_store.py
@@ -411,19 +411,15 @@ class CoinStore:
         await cursor.close()
 
     # Update coin_record to be spent in DB
-    async def _set_spent(self, coin_name: bytes32, index: uint32) -> CoinRecord:
-        current: Optional[CoinRecord] = await self.get_coin_record(coin_name)
-        if current is None:
-            raise ValueError(f"Cannot spend a coin that does not exist in db: {coin_name}")
+    async def _set_spent(self, coin_name: bytes32, index: uint32):
 
-        assert not current.spent  # Redundant sanity check, already checked in block_body_validation
-        spent: CoinRecord = CoinRecord(
-            current.coin,
-            current.confirmed_block_index,
-            index,
-            True,
-            current.coinbase,
-            current.timestamp,
-        )  # type: ignore # noqa
-        await self._add_coin_record(spent, True)
-        return spent
+        # if this coin is in the cache, mark it as spent in there
+        r = self.coin_record_cache.get(coin_name)
+        if r is not None:
+            self.coin_record_cache.put(
+                r.name, CoinRecord(r.coin, r.confirmed_block_index, index, True, r.coinbase, r.timestamp)
+            )
+
+        await self.coin_record_db.execute(
+            f"UPDATE OR FAIL coin_record SET spent=1,spent_index=? WHERE coin_name=?", (index, coin_name.hex())
+        )