--- a/tests/core/full_node/test_coin_store.py
+++ b/tests/core/full_node/test_coin_store.py
@@ -194,8 +194,7 @@ class TestCoinStoreWithBlocks:
                     coins = block.get_included_reward_coins()
                     records = [await coin_store.get_coin_record(coin.name()) for coin in coins]
 
-                    for record in records:
-                        await coin_store._set_spent(record.coin.name(), block.height)
+                    await coin_store._set_spent([r.name for r in records], block.height)
 
                     records = [await coin_store.get_coin_record(coin.name()) for coin in coins]
                     for record in records:
@@ -210,7 +209,7 @@ class TestCoinStoreWithBlocks:
         async with DBConnection() as db_wrapper:
             coin_store = await CoinStore.create(db_wrapper, cache_size=uint32(cache_size))
 
-            records: List[Optional[CoinRecord]] = []
+            records: List[CoinRecord] = []
 
             for block in blocks:
                 if block.is_transaction_block():
@@ -230,9 +229,7 @@ class TestCoinStoreWithBlocks:
                     coins = block.get_included_reward_coins()
                     records = [await coin_store.get_coin_record(coin.name()) for coin in coins]
 
-                    for record in records:
-                        assert record is not None
-                        await coin_store._set_spent(record.coin.name(), block.height)
+                    await coin_store._set_spent([r.name for r in records], block.height)
 
                     records = [await coin_store.get_coin_record(coin.name()) for coin in coins]
                     for record in records: