Please review commits one at a time.

The first commit adds a benchmark program for the `CoinStore` class. Subsequent commits improve its performance. Some commits may be controversial, such as the one changing `syncronous` to `OFF`. Let's discuss which ones to keep and which ones to drop.

These are the results from this benchmark, running under Ubuntu on a decent desktop: Intel(R) Xeon(R) CPU E3-1245 v3 @ 3.40GHz woth 16 GB of RAM. The main feature is that it has a spinning disk. Specifically, the benchmark ran with the database on the spinning disk.

**EDIT**

This patch was rebased onto main, changing the return value from `new_block()` as well as making the benchmark larger. It now creates a larger database before starting the test. The idea is to increase fidelity to a realistic workflow. It certainly made the test take longer to run, but is probably still not large enough to fully reflect realistic behavior.

## baseline

```
MOSTLY ADDITIONS, time: 1779.4565s additions: 400400 removals: 20000
MOSTLY REMOVALS, time: 1688.2673s additions: 600 removals: 140000
FULLBLOCKS, time: 3486.4196s additions: 400400 removals: 400000
```

~57 tx/second*


## synchronous=NORMAL

```
MOSTLY ADDITIONS, time: 1633.3031s additions: 400400 removals: 20000
MOSTLY REMOVALS, time: 965.8632s additions: 600 removals: 140000
FULLBLOCKS, time: 2449.2953s additions: 400400 removals: 400000
```

~81 tx/second*


## synchronous=OFF

```
MOSTLY ADDITIONS, time: 50.5357s additions: 400400 removals: 20000
MOSTLY REMOVALS, time: 6.5864s additions: 600 removals: 140000
FULLBLOCKS, time: 79.4324s additions: 400400 removals: 400000
```

~2531 tx/second*


\* If one transaction is two inputs and two outputs



## OUTDATED BENCHMARKS

> 
> Initial commit:
> 
> ```
> MOSTLY ADDITIONS, time: 84.07s additions: 398000 removals: 19900
> MOSTLY REMOVALS, time: 956.1s additions: 200 removals: 140000
> FULLBLOCKS, time: 2030s additions: 400000 removals: 400000
> ```
> 
> _set_spent() optimization
> 
> ```
> MOSTLY ADDITIONS, time: 74.3488s additions: 398000 removals: 19900
> MOSTLY REMOVALS, time: 19.0282s additions: 200 removals: 140000
> FULLBLOCKS, time: 1336.4314s additions: 400000 removals: 400000
> ```
> 
> synchronous=OFF
> 
> ```
> MOSTLY ADDITIONS, time: 74.9006s additions: 398000 removals: 19900
> MOSTLY REMOVALS, time: 8.7613s additions: 200 removals: 140000
> FULLBLOCKS, time: 109.0237s additions: 400000 removals: 400000
> ```
> 
> executemany (inserts):
> 
> ```
> MOSTLY ADDITIONS, time: 22.5494s additions: 398000 removals: 19900
> MOSTLY REMOVALS, time: 10.2330s additions: 200 removals: 140000
> FULLBLOCKS, time: 59.9542s additions: 400000 removals: 400000
> ```
> 
> executemany (_set_spent):
> 
> ```
> MOSTLY ADDITIONS, time: 24.6169s additions: 398000 removals: 19900
> MOSTLY REMOVALS, time: 8.9951s additions: 200 removals: 140000
> FULLBLOCKS, time: 67.0182s additions: 400000 removals: 400000
> ```

These numbers should be taken with a grain of salt as they can be quite noisy. Also, during these tests, I suspect the database would still fit in working memory, which further might skew this result, compared to a real-life scenario where the DB is a lot larger.